chore(#279): resolve duplicate test numbers in tests/
- Merge tests/18-blind-spots-2.test.js content into tests/18-blind-spots-round2.test.js
  (former 18-2 file deleted; 18-round2 already had a header note for this merge)
- Rename tests/35-einheit-groesse-calc.test.js -> tests/46-einheit-groesse-calc.test.js
  (resolves duplicate 35; 35-carryover-isTabDone-fix.test.js keeps its number)
- Gaps 20, 25, 31, 32 left intact per task spec (no renumbering, no filling)

All 782 vitest tests pass; ESLint clean.
